fix: slightly better banner scroll behavior
