+++ apps/mgmt-ui/src/pages/_app.tsx
@@ -1,13 +1,11 @@
 import { SupaglueHead } from '@/components/SupaglueHead';
 import { NotificationManager } from '@/context/notification';
-import { useActiveApplication } from '@/hooks/useActiveApplication';
 import Navigator from '@/layout/Navigator';
 import '@/styles/globals.css';
 import { ClerkProvider, RedirectToSignIn, SignedIn, SignedOut, useUser } from '@clerk/nextjs';
 import { Box, CssBaseline, StyledEngineProvider, useMediaQuery } from '@mui/material';
 import { createTheme, ThemeProvider } from '@mui/material/styles';
 import { LicenseInfo } from '@mui/x-license-pro';
-import { DateTime } from 'luxon';
 import { SessionProvider } from 'next-auth/react';
 import type { AppProps } from 'next/app';
 import { useRouter } from 'next/router';
@@ -32,8 +30,8 @@ if (typeof window !== 'undefined') {
     api_host: process.env.FRONTEND_URL
       ? `${process.env.FRONTEND_URL}/ingest`
       : window.location.port === '3000'
-        ? `http://${window.location.hostname}:3000/ingest`
-        : `https://${window.location.hostname}/ingest`,
+      ? `http://${window.location.hostname}:3000/ingest`
+      : `https://${window.location.hostname}/ingest`,
   });
 }
 
@@ -271,7 +269,6 @@ const PosthogClerkUserIdentifier = (props: { children: ReactNode }) => {
 };
 
 function InnerApp(props: { signedIn: boolean; children: ReactNode } & SupaglueProps) {
-  const { activeApplication } = useActiveApplication();
   const { signedIn, children } = props;
   const [mobileOpen, setMobileOpen] = useState(false);
   const isSmUp = useMediaQuery(theme.breakpoints.up('sm'));
@@ -282,7 +279,6 @@ function InnerApp(props: { signedIn: boolean; children: ReactNode } & SupagluePr
 
   return (
     <>
-      {signedIn && activeApplication && !activeApplication.isPaid && <Banner createdAt={activeApplication.createdAt} />}
       <Box sx={{ display: 'flex', minHeight: '100vh' }}>
         <CssBaseline />
         {/* Don't show navigator if not signed in */}
@@ -290,7 +286,7 @@ function InnerApp(props: { signedIn: boolean; children: ReactNode } & SupagluePr
           <Box component="nav" sx={{ width: { sm: drawerWidth }, flexShrink: { sm: 0 } }}>
             {isSmUp ? null : (
               <Navigator
-                PaperProps={{ style: { width: drawerWidth, top: 'unset' } }}
+                PaperProps={{ style: { width: drawerWidth } }}
                 variant="temporary"
                 open={mobileOpen}
                 onClose={handleDrawerToggle}
@@ -298,7 +294,7 @@ function InnerApp(props: { signedIn: boolean; children: ReactNode } & SupagluePr
               />
             )}
             <Navigator
-              PaperProps={{ style: { width: drawerWidth, top: 'unset' } }}
+              PaperProps={{ style: { width: drawerWidth } }}
               sx={{ display: { sm: 'block', xs: 'none' } }}
               {...props}
             />
@@ -308,59 +304,5 @@ function InnerApp(props: { signedIn: boolean; children: ReactNode } & SupagluePr
         {children}
       </Box>
     </>
-  );
-}
-
-function Banner(props: { createdAt: string }) {
-  const { createdAt } = props;
-
-  const createdAtDate = DateTime.fromJSDate(new Date(createdAt));
-  const nowDate = DateTime.fromJSDate(new Date());
-
-  const dateDiff = nowDate.diff(createdAtDate, ['days']).toObject();
-  const daysRemaining = Math.max(0, 30 - Math.floor(dateDiff.days ?? 30));
-
-  return (
-    <div className="relative isolate flex items-center gap-x-6 overflow-hidden bg-gray-50 px-6 py-2.5 sm:px-3.5 sm:before:flex-1 z-[1300]">
-      <div
-        className="absolute left-[max(-7rem,calc(50%-52rem))] top-1/2 -z-10 -translate-y-1/2 transform-gpu blur-2xl"
-        aria-hidden="true"
-      >
-        <div
-          className="aspect-[577/310] w-[36.0625rem] bg-gradient-to-r from-[#ff80b5] to-[#9089fc] opacity-30"
-          style={{
-            clipPath:
-              'polygon(74.8% 41.9%, 97.2% 73.2%, 100% 34.9%, 92.5% 0.4%, 87.5% 0%, 75% 28.6%, 58.5% 54.6%, 50.1% 56.8%, 46.9% 44%, 48.3% 17.4%, 24.7% 53.9%, 0% 27.9%, 11.9% 74.2%, 24.9% 54.1%, 68.6% 100%, 74.8% 41.9%)',
-          }}
-        />
-      </div>
-      <div
-        className="absolute left-[max(45rem,calc(50%+8rem))] top-1/2 -z-10 -translate-y-1/2 transform-gpu blur-2xl"
-        aria-hidden="true"
-      >
-        <div
-          className="aspect-[577/310] w-[36.0625rem] bg-gradient-to-r from-[#ff80b5] to-[#9089fc] opacity-30"
-          style={{
-            clipPath:
-              'polygon(74.8% 41.9%, 97.2% 73.2%, 100% 34.9%, 92.5% 0.4%, 87.5% 0%, 75% 28.6%, 58.5% 54.6%, 50.1% 56.8%, 46.9% 44%, 48.3% 17.4%, 24.7% 53.9%, 0% 27.9%, 11.9% 74.2%, 24.9% 54.1%, 68.6% 100%, 74.8% 41.9%)',
-          }}
-        />
-      </div>
-      <div className="flex flex-wrap items-center gap-x-4 gap-y-2">
-        <p className="text-sm leading-6 text-gray-900">
-          <strong className="font-semibold">Free 30 day trial</strong>
-          <svg viewBox="0 0 2 2" className="mx-2 inline h-0.5 w-0.5 fill-current" aria-hidden="true">
-            <circle cx={1} cy={1} r={1} />
-          </svg>
-          There are <span className="font-semibold">{daysRemaining}</span> days remaining in your free Supaglue trial.
-        </p>
-        <a
-          href="mailto:support@supaglue.com"
-          className="flex-none rounded-full bg-gray-900 px-3.5 py-1 text-sm font-semibold text-white shadow-sm hover:bg-gray-700 focus-visible:outline focus-visible:outline-2 focus-visible:outline-offset-2 focus-visible:outline-gray-900"
-        >
-          Upgrade now<span aria-hidden="true">&rarr;</span>
-        </a>
-      </div>
-    </div>
   );
 }
